flask-template/app.py
from flask import Flask, render_template, request
import json

app = Flask(__name__)

@app.route('/')
def index():
    return render_template('airlineweb_dashboard.html')

@app.route('/results', methods=['POST'])
def get_insights_route():
    filters = request.form
    data = scrape_airline_data()  # Scrape data from the airline website
    insights = get_insights(data)
    return render_template('results.html', insights=insights)

def scrape_airline_data():
    # Dummy implementation for demonstration; replace with actual scraping logic
    return {"flights": [], "message": "Sample data from scrape_airline_data"}

def get_insights(data):
    # Add your logic here to process 'data' and return insights
    # For demonstration, just return the data
    return data
# Removed unused 'results' function to avoid confusion and errors
    #insights = get_insights(data)  # Solve the Error
    insights = get_insights(data)  
    
    # Handle insights and render them in the template
    # For demonstration, we assume insights is a dictionary
    
    return render_template('results.html', insights=insights)


if __name__ == '__main__':
    # Removed invalid example usage of get_insights() without arguments
    app.run(debug=True)
